+++ packages/web3-core/types/index.d.ts
@@ -138,6 +138,11 @@ export interface Transaction {
     maxFeePerGas?: number | string | BN;
     gas: number;
     input: string;
+    chainId?: string;
+    accessList?: AccessList;
+    v?: string;
+    r?: string;
+    s?: string;
 }
 
 export interface TransactionConfig {
@@ -456,3 +461,10 @@ export type provider =
     | AbstractProvider
     | string
     | null;
+
+export interface AccessTuple {
+    address: string;
+    storageKeys: string[];
+}
+
+export type AccessList = AccessTuple[];