+++ packages/web3-eth-iban/types/tests/iban-tests.ts
@@ -20,10 +20,13 @@
  * @date 2018
  */
 
-import { Iban, IndirectOptions } from 'web3-eth-iban';
+import IbanDefault, { Iban, IndirectOptions } from 'web3-eth-iban';
 
 const iban = 'XE7338O073KYGTWWZN0F2WZ0R8PX5ZPPZS';
 const address = '0x45cd08334aeedd8a06265b2ae302e3597d8faa28';
+
+// $ExpectType Iban
+const iban_default = new IbanDefault(iban);
 
 // $ExpectType Iban
 const ibanClass = new Iban(iban);