+++ packages/web3-eth/types/index.d.ts
@@ -36,7 +36,8 @@ import {
     chain,
     BlockNumber,
     LogsOptions,
-    PastLogsOptions
+    PastLogsOptions,
+    AccessTuple
 } from 'web3-core';
 import {Subscription} from 'web3-core-subscriptions';
 import {AbiCoder} from 'web3-eth-abi';
@@ -452,11 +453,6 @@ export interface BlockTransactionString extends BlockTransactionBase {
     transactions: string[];
 }
 
-export interface AccessTuple {
-    address: string;
-    storageKeys: string[];
-}
-
 export interface CreateAccessList {
     accessList: AccessTuple[];
     error?: string;
@@ -485,3 +481,5 @@ export interface FeeHistoryResult {
     oldestBlock: number;
     reward: string[][];
 }
+
+export default Eth