+++ scripts/eth2.sh
@@ -21,7 +21,7 @@ echo "ETH2: Lighthouse Node (requires docker)    "
 echo ">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>"
 echo " "
 
-nohup docker run -p 9000:9000 -p 127.0.0.1:5052:5052 -v ${HOME}/.lighthouse:/root/.lighthouse1 sigp/lighthouse:v2.1.5 lighthouse --network mainnet beacon --http --http-address 0.0.0.0 > /dev/null 2>&1 &
+nohup docker run -p 9000:9000 -p 127.0.0.1:5052:5052 -v ${HOME}/.lighthouse:/root/.lighthouse1 sigp/lighthouse:latest lighthouse --network mainnet beacon --http --http-address 0.0.0.0 > /dev/null 2>&1 &
 
 npm run build